requirements.txt
django
yfinance
sktime
python-dotenv
pandas
scikit-learn
deta
pandas-ta
plotly